acomodando funciones en triggers

+++ src/DatabaseScripts/Functions/NativeFunctions/202005122223_convertSecondsInMinutes.sql
@@ -10,7 +10,7 @@ drop function if exists convertSecondsInMinutes;
 delimiter //
 create function convertSecondsInMinutes
 (
-pSeconds int
+pSeconds float
 )
 returns float
 not deterministic
